port create_ticket to twig and fix bug with ticket_history
