+++ src/public/controllers/tickets/ticket_history.php
@@ -1,7 +1,7 @@
 <?php
 require_once('helpdbconnect.php');
-require_once(from_root("/includes/block_file.php"));
-require_once(from_root("/includes/ticket_utils.php"));
+require_once("block_file.php");
+require_once("ticket_utils.php");
 require from_root("/../vendor/autoload.php");
 require from_root("/new-controllers/ticket_base_variables.php");
